feat(fleet-console): replace topbar with status bar shell
