+++ runtime/fleet-console/core/client/src/styles/layout.css
@@ -1,28 +1,26 @@
 .console-shell {
+  --console-statusbar-height: 28px;
   display: grid;
-  grid-template-rows: 64px minmax(0, 1fr);
+  grid-template-rows: minmax(0, 1fr) var(--console-statusbar-height);
   height: 100%;
   padding: var(--space-4);
-  gap: var(--space-4);
+  gap: var(--space-3);
   transition:
     grid-template-rows var(--duration-slow) var(--ease-spring),
     gap var(--duration-slow) var(--ease-spring),
     padding var(--duration-slow) var(--ease-spring);
 }
 
-/* 맵 최대화: GNB(topbar) 행을 0으로 접고 셸 여백을 없애 캔버스를 전체로 채운다(app.tsx가 Map canvas 화면에서만 부여). */
 .console-shell.is-map-fullscreen {
-  grid-template-rows: 0 minmax(0, 1fr);
+  grid-template-rows: minmax(0, 1fr) 0;
   gap: 0;
   padding: 0;
 }
 
-.console-shell.is-map-fullscreen .topbar {
-  /* codex-rise 첫-페인트 애니메이션이 both fill로 opacity/transform 끝상태를 잡고 있어
-     이를 해제해야 아래 fade+slide가 .topbar transition을 타고 적용된다. */
+.console-shell.is-map-fullscreen .statusbar {
   animation: none;
   opacity: 0;
-  transform: translateY(-12px);
+  transform: translateY(12px);
   pointer-events: none;
   overflow: hidden;
 }
@@ -87,12 +85,12 @@
   animation-delay: 200ms;
 }
 
-.topbar,
 .stage {
   animation: codex-rise 720ms var(--ease-spring) both;
 }
 
-.topbar {
+.statusbar {
+  animation: codex-rise 720ms var(--ease-spring) both;
   animation-delay: 40ms;
   transition:
     opacity var(--duration-base) ease,
@@ -115,32 +113,15 @@
   }
 }
 
-/* 우현 오버레이(Side 패널·scrim·검색 팔레트)의 상단 오프셋 — topbar bottom에 맞춰 GNB를 덮지 않게 한다.
-   기본은 1행 topbar(64px) + 셸 여백. 좁은 폭에서 topbar가 2행으로 stack되면 아래 breakpoint가 갱신한다. */
-:root {
-  --console-topbar-offset: calc(64px + var(--space-4));
-}
-
-/* GNB가 2행으로 stack되는 폭(components.css의 ≤1080px ladder)에서는 고정 64px topbar 행으로는
-   부족하므로, topbar 행을 내용 높이에 맞춰 본문으로 흘러내리지 않게 한다. */
 @media (max-width: 1080px) {
-  :root {
-    /* 2행 stack된 topbar 실제 높이(≈108px) + 좁은 폭 셸 padding(space-3). */
-    --console-topbar-offset: calc(108px + var(--space-3));
-  }
-
   .console-shell {
-    grid-template-rows: auto minmax(0, 1fr);
+    grid-template-rows: minmax(0, 1fr) var(--console-statusbar-height);
   }
 }
 
 @media (max-width: 760px) {
-  :root {
-    --console-topbar-offset: calc(108px + var(--space-2));
-  }
-
   .console-shell {
-    grid-template-rows: auto minmax(0, 1fr);
+    grid-template-rows: minmax(0, 1fr) var(--console-statusbar-height);
     padding: var(--space-2);
   }
